Cleaned up after new list functionality.
Some adjustments after code inspection.
