+++ src/main/kotlin/no/nav/klage/search/api/controller/BehandlingSearchController.kt
@@ -7,11 +7,8 @@ import no.nav.klage.search.api.mapper.BehandlingerSearchCriteriaMapper
 import no.nav.klage.search.api.view.*
 import no.nav.klage.search.config.SecurityConfiguration.Companion.ISSUER_AAD
 import no.nav.klage.search.domain.personsoek.PersonSearchResponse
-import no.nav.klage.search.domain.saksbehandler.EnhetMedLovligeYtelser
 import no.nav.klage.search.exceptions.PersonNotFoundException
 import no.nav.klage.search.service.PersonSearchService
-import no.nav.klage.search.service.saksbehandler.InnloggetSaksbehandlerService
-import no.nav.klage.search.service.saksbehandler.OAuthTokenService
 import no.nav.klage.search.util.getLogger
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.web.bind.annotation.PostMapping
@@ -26,8 +23,6 @@ import org.springframework.web.bind.annotation.RestController
 class BehandlingSearchController(
     private val behandlingListMapper: BehandlingListMapper,
     private val behandlingerSearchCriteriaMapper: BehandlingerSearchCriteriaMapper,
-    private val oAuthTokenService: OAuthTokenService,
-    private val innloggetSaksbehandlerService: InnloggetSaksbehandlerService,
     private val personSearchService: PersonSearchService,
 ) {
 
@@ -71,46 +66,5 @@ class BehandlingSearchController(
             }
         )
     }
-
-    // Not in use atm
-/*    @PostMapping("/relaterte")
-    fun getRelaterteKlagebehandlinger(
-        @RequestBody input: SearchPersonByFnrInput
-    ): KlagebehandlingerListRespons {
-        //TODO: Move logic to PersonsoekService
-        val lovligeTemaer = getEnhetOrThrowException(input.enhet).temaer
-        val sivilstand: Sivilstand? = pdlFacade.getPersonInfo(input.query).sivilstand
-
-        val searchCriteria = KlagebehandlingerSearchCriteria(
-            statuskategori = KlagebehandlingerSearchCriteria.Statuskategori.ALLE,
-            ferdigstiltFom = LocalDate.now().minusMonths(12),
-            foedselsnr = listOf(input.query),
-            extraPersonAndTema = sivilstand?.let {
-                KlagebehandlingerSearchCriteria.ExtraPersonAndTema(
-                    foedselsnr = it.foedselsnr,
-                    temaer = TemaTilgjengeligeForEktefelle.temaerTilgjengeligForEktefelle(environment).toList()
-                )
-            },
-            offset = 0,
-            limit = 100,
-            projection = KlagebehandlingerSearchCriteria.Projection.UTVIDET,
-        )
-
-        val esResponse = elasticsearchService.findByCriteria(searchCriteria)
-        return KlagebehandlingerListRespons(
-            antallTreffTotalt = esResponse.totalHits.toInt(),
-            klagebehandlinger = klagebehandlingListMapper.mapEsKlagebehandlingerToListView(
-                esKlagebehandlinger = esResponse.searchHits.map { it.content },
-                viseUtvidet = true,
-                viseFullfoerte = true,
-                saksbehandler = searchCriteria.saksbehandler,
-                tilgangTilTemaer = lovligeTemaer,
-                sivilstand = sivilstand
-            )
-        )
-    }*/
-
-
-
 }
