+++ src/main/kotlin/no/nav/klage/search/api/controller/OppgaverListController.kt
@@ -9,7 +9,6 @@ import no.nav.klage.search.api.mapper.BehandlingerSearchCriteriaMapper
 import no.nav.klage.search.api.view.*
 import no.nav.klage.search.config.SecurityConfiguration.Companion.ISSUER_AAD
 import no.nav.klage.search.exceptions.MissingTilgangException
-import no.nav.klage.search.exceptions.NotMatchingUserException
 import no.nav.klage.search.service.ElasticsearchService
 import no.nav.klage.search.service.saksbehandler.InnloggetSaksbehandlerService
 import no.nav.klage.search.service.saksbehandler.OAuthTokenService
@@ -39,91 +38,124 @@ class OppgaverListController(
         summary = "Hent ledige oppgaver for en saksbehandler",
         description = "Henter alle ledige oppgaver saksbehandler har tilgang til."
     )
-    @GetMapping("/ansatte/{navIdent}/oppgaver/ledige", produces = ["application/json"])
+    @GetMapping("/oppgaver/ledige", produces = ["application/json"])
     fun getMineLedigeOppgaver(
         queryParams: MineLedigeOppgaverQueryParams
-    ): BehandlingerListRespons {
+    ): BehandlingerListResponse {
         logger.debug("Params: {}", queryParams)
 
         val ytelser = getYtelserQueryListForSaksbehandler(
-            queryParams = queryParams,
+            ytelser = queryParams.ytelser,
             tildelteYtelser = innloggetSaksbehandlerService.getTildelteYtelserForSaksbehandler()
         )
         //TODO: Dette hadde vært bedre å håndtere i ElasticsearchService enn her
         if (ytelser.isEmpty()) {
-            return emptyResponse()
+            return BehandlingerListResponse(antallTreffTotalt = 0, behandlinger = emptyList())
         }
 
-        //val hjemler: List<String> = lovligeValgteHjemler(queryParams = queryParams, ytelser = ytelser)
         val searchCriteria = behandlingerSearchCriteriaMapper.toLedigeOppgaverSearchCriteria(
-            queryParams = queryParams.copy(ytelser = ytelser) //, hjemler = hjemler),
+            queryParams = queryParams.copy(ytelser = ytelser),
         )
 
         val esResponse = elasticsearchService.findLedigeOppgaverByCriteria(searchCriteria)
-        return BehandlingerListRespons(
+        return BehandlingerListResponse(
             antallTreffTotalt = esResponse.totalHits.toInt(),
             behandlinger = behandlingListMapper.mapEsBehandlingerToListView(
                 esBehandlinger = esResponse.searchHits.map { it.content },
-                visePersonData = false,
             )
         )
+    }
+
+    @Operation(
+        summary = "Hent oppgave",
+        description = "Hent oppgave."
+    )
+    @GetMapping("/oppgaver/{behandlingId}", produces = ["application/json"])
+    fun getOppgave(
+        @PathVariable behandlingId: String,
+    ): BehandlingView {
+        logger.debug("getOppgave: {}", behandlingId)
+
+        val searchCriteria = behandlingerSearchCriteriaMapper.toBehandlingIdSearchCriteria(
+            behandlingId = behandlingId,
+        )
+        val esResponse = elasticsearchService.findOppgaveByBehandlingId(searchCriteria)
+
+        val result = behandlingListMapper.mapEsBehandlingerToBehandlingView(
+            esBehandlinger = esResponse.searchHits.map { it.content },
+            visePersonData = false,
+        )
+
+        return result.first()
+    }
+
+    @Operation(
+        summary = "Hent person",
+        description = "Hent person."
+    )
+    @GetMapping("/oppgaver/{behandlingId}/person", produces = ["application/json"])
+    fun getPersonByOppgave(
+        @PathVariable behandlingId: String,
+    ): PersonView {
+        logger.debug("getPersonByOppgave: {}", behandlingId)
+
+        val searchCriteria = behandlingerSearchCriteriaMapper.toBehandlingIdSearchCriteria(
+            behandlingId = behandlingId,
+        )
+        val esResponse = elasticsearchService.findOppgaveByBehandlingId(searchCriteria)
+
+        val result = behandlingListMapper.mapEsBehandlingerToBehandlingView(
+            esBehandlinger = esResponse.searchHits.map { it.content },
+            visePersonData = true,
+        )
+
+        return result.first().person!!
     }
 
     @Operation(
         summary = "Hent ferdigstilte oppgaver for en ansatt",
         description = "Henter alle ferdigstilte oppgaver som saksbehandler har tilgang til."
     )
-    @GetMapping("/ansatte/{navIdent}/oppgaver/ferdigstilte", produces = ["application/json"])
+    @GetMapping("/oppgaver/ferdigstilte", produces = ["application/json"])
     fun getMineFerdigstilteOppgaver(
-        @Parameter(name = "NavIdent til en ansatt")
-        @PathVariable navIdent: String,
         queryParams: MineFerdigstilteOppgaverQueryParams
-    ): BehandlingerListRespons {
+    ): BehandlingerListResponse {
         logger.debug("Params: {}", queryParams)
-        validateNavIdent(navIdent)
 
         val searchCriteria = behandlingerSearchCriteriaMapper.toSaksbehandlersFerdigstilteOppgaverSearchCriteria(
-            navIdent = navIdent,
+            navIdent = oAuthTokenService.getInnloggetIdent(),
             queryParams = queryParams
         )
 
         val esResponse = elasticsearchService.findSaksbehandlersFerdigstilteOppgaverByCriteria(searchCriteria)
-        return BehandlingerListRespons(
+        return BehandlingerListResponse(
             antallTreffTotalt = esResponse.totalHits.toInt(),
             behandlinger = behandlingListMapper.mapEsBehandlingerToListView(
                 esBehandlinger = esResponse.searchHits.map { it.content },
-                visePersonData = true,
             ),
         )
     }
-
-    private fun emptyResponse(): BehandlingerListRespons =
-        BehandlingerListRespons(antallTreffTotalt = 0, behandlinger = emptyList())
 
     @Operation(
         summary = "Hent uferdige oppgaver for en ansatt",
         description = "Henter alle uferdige oppgaver som saksbehandler har tilgang til."
     )
-    @GetMapping("/ansatte/{navIdent}/oppgaver/uferdige", produces = ["application/json"])
+    @GetMapping("/oppgaver/uferdige", produces = ["application/json"])
     fun getMineUferdigeOppgaver(
-        @Parameter(name = "NavIdent til en ansatt")
-        @PathVariable navIdent: String,
         queryParams: MineUferdigeOppgaverQueryParams
-    ): BehandlingerListRespons {
+    ): BehandlingerListResponse {
         logger.debug("Params: {}", queryParams)
-        validateNavIdent(navIdent)
 
         val searchCriteria = behandlingerSearchCriteriaMapper.toSaksbehandlersUferdigeOppgaverSearchCriteria(
-            navIdent = navIdent,
+            navIdent = oAuthTokenService.getInnloggetIdent(),
             queryParams = queryParams
         )
 
         val esResponse = elasticsearchService.findSaksbehandlersUferdigeOppgaverByCriteria(searchCriteria)
-        return BehandlingerListRespons(
+        return BehandlingerListResponse(
             antallTreffTotalt = esResponse.totalHits.toInt(),
             behandlinger = behandlingListMapper.mapEsBehandlingerToListView(
                 esBehandlinger = esResponse.searchHits.map { it.content },
-                visePersonData = true,
             ),
         )
     }
@@ -132,45 +164,128 @@ class OppgaverListController(
         summary = "Hent oppgaver satt på vent for en ansatt",
         description = "Henter alle oppgaver satt på vent som saksbehandler har tilgang til."
     )
-    @GetMapping("/ansatte/{navIdent}/oppgaver/paavent", produces = ["application/json"])
+    @GetMapping("/oppgaver/paavent", produces = ["application/json"])
     fun getMineOppgaverPaaVent(
-        @Parameter(name = "NavIdent til en ansatt")
-        @PathVariable navIdent: String,
         queryParams: MineOppgaverPaaVentQueryParams
-    ): BehandlingerListRespons {
+    ): BehandlingerListResponse {
         logger.debug("Params: {}", queryParams)
-        validateNavIdent(navIdent)
 
         val searchCriteria = behandlingerSearchCriteriaMapper.toSaksbehandlersOppgaverPaaVentSearchCriteria(
-            navIdent = navIdent,
+            navIdent = oAuthTokenService.getInnloggetIdent(),
             queryParams = queryParams
         )
 
         val esResponse = elasticsearchService.findSaksbehandlersOppgaverPaaVentByCriteria(searchCriteria)
-        return BehandlingerListRespons(
+        return BehandlingerListResponse(
             antallTreffTotalt = esResponse.totalHits.toInt(),
             behandlinger = behandlingListMapper.mapEsBehandlingerToListView(
                 esBehandlinger = esResponse.searchHits.map { it.content },
-                visePersonData = true,
             ),
         )
     }
 
     @Operation(
-        summary = "Hent antall utildelte behandlinger for tilgjengelig for saksbehandler der fristen gått ut",
-        description = "Hent antall utildelte behandlinger for tilgjengelig for saksbehandler der fristen gått ut"
+        summary = "Hent enhetens ferdigstilte oppgaver",
+        description = "Henter alle ferdigstilte oppgaver for enheten som saksbehandler har tilgang til."
     )
-    @GetMapping("/ansatte/{navIdent}/antalloppgavermedutgaattefrister", produces = ["application/json"])
+    @GetMapping(
+        "/enhet/{enhetId}/oppgaver/tildelte/ferdigstilte",
+        produces = ["application/json"]
+    )
+    fun getEnhetensFerdigstilteOppgaver(
+        @Parameter(name = "EnhetId til enheten den ansatte jobber i")
+        @PathVariable enhetId: String,
+        queryParams: EnhetensFerdigstilteOppgaverQueryParams
+    ): BehandlingerListResponse {
+        logger.debug("Params: {}", queryParams)
+        validateRettigheterForEnhetensTildelteOppgaver()
+
+        val searchCriteria = behandlingerSearchCriteriaMapper.toEnhetensFerdigstilteOppgaverSearchCriteria(
+            enhetId = enhetId,
+            queryParams = queryParams
+        )
+
+        val esResponse = elasticsearchService.findEnhetensFerdigstilteOppgaverByCriteria(searchCriteria)
+        return BehandlingerListResponse(
+            antallTreffTotalt = esResponse.totalHits.toInt(),
+            behandlinger = behandlingListMapper.mapAnonymeEsBehandlingerToListView(
+                esBehandlinger = esResponse.searchHits.map { it.content },
+            ),
+        )
+    }
+
+    @Operation(
+        summary = "Hent enhetens oppgaver på vent",
+        description = "Henter alle oppgaver satt på vent for enheten som saksbehandler har tilgang til."
+    )
+    @GetMapping(
+        "/enhet/{enhetId}/oppgaver/tildelte/paavent",
+        produces = ["application/json"]
+    )
+    fun getEnhetensOppgaverPaaVent(
+        @Parameter(name = "EnhetId til enheten den ansatte jobber i")
+        @PathVariable enhetId: String,
+        queryParams: EnhetensOppgaverPaaVentQueryParams
+    ): BehandlingerListResponse {
+        logger.debug("Params: {}", queryParams)
+        validateRettigheterForEnhetensTildelteOppgaver()
+
+        val searchCriteria = behandlingerSearchCriteriaMapper.toEnhetensOppgaverPaaVentSearchCriteria(
+            enhetId = enhetId,
+            queryParams = queryParams
+        )
+
+        val esResponse = elasticsearchService.findEnhetensOppgaverPaaVentByCriteria(searchCriteria)
+        return BehandlingerListResponse(
+            antallTreffTotalt = esResponse.totalHits.toInt(),
+            behandlinger = behandlingListMapper.mapAnonymeEsBehandlingerToListView(
+                esBehandlinger = esResponse.searchHits.map { it.content },
+            ),
+        )
+    }
+
+    @Operation(
+        summary = "Hent uferdige oppgaver for en enhet",
+        description = "Henter alle uferdige oppgaver i enheten som saksbehandler har tilgang til."
+    )
+    @GetMapping(
+        "/enhet/{enhetId}/oppgaver/tildelte/uferdige",
+        produces = ["application/json"]
+    )
+    fun getEnhetensUferdigeOppgaver(
+        @Parameter(name = "EnhetId til enheten den ansatte jobber i")
+        @PathVariable enhetId: String,
+        queryParams: EnhetensUferdigeOppgaverQueryParams
+    ): BehandlingerListResponse {
+        logger.debug("Params: {}", queryParams)
+        validateRettigheterForEnhetensTildelteOppgaver()
+
+        val searchCriteria = behandlingerSearchCriteriaMapper.toEnhetensUferdigeOppgaverSearchCriteria(
+            enhetId = enhetId,
+            queryParams = queryParams
+        )
+
+        val esResponse = elasticsearchService.findEnhetensUferdigeOppgaverByCriteria(searchCriteria)
+        return BehandlingerListResponse(
+            antallTreffTotalt = esResponse.totalHits.toInt(),
+            behandlinger = behandlingListMapper.mapAnonymeEsBehandlingerToListView(
+                esBehandlinger = esResponse.searchHits.map { it.content },
+            ),
+        )
+    }
+
+    @Operation(
+        summary = "Hent antall utildelte behandlinger tilgjengelig for saksbehandler der fristen gått ut",
+        description = "Hent antall utildelte behandlinger tilgjengelig for saksbehandler der fristen gått ut"
+    )
+    @GetMapping("/antalloppgavermedutgaattefrister", produces = ["application/json"])
     fun getUtgaatteFristerAvailableToSaksbehandlerCount(
-        @Parameter(name = "NavIdent til en ansatt")
-        @PathVariable navIdent: String,
-        queryParams: MineLedigeOppgaverQueryParams
+        queryParams: MineLedigeOppgaverCountQueryParams
     ): AntallUtgaatteFristerResponse {
         logger.debug("Params: {}", queryParams)
-        validateNavIdent(navIdent)
 
         val ytelser = getYtelserQueryListForSaksbehandler(
-            queryParams = queryParams,
+            ytelser = queryParams.ytelser,
             tildelteYtelser = innloggetSaksbehandlerService.getTildelteYtelserForSaksbehandler()
         )
         //TODO: Dette hadde vært bedre å håndtere i ElasticsearchService enn her
@@ -181,102 +296,10 @@ class OppgaverListController(
         return AntallUtgaatteFristerResponse(
             antall = elasticsearchService.countLedigeOppgaverMedUtgaattFristByCriteria(
                 criteria = behandlingerSearchCriteriaMapper.toSearchCriteriaForLedigeMedUtgaattFrist(
-                    navIdent = navIdent,
-                    queryParams = queryParams.copy(ytelser = ytelser) //, hjemler = hjemler),
+                    queryParams = queryParams.copy(ytelser = ytelser),
                 )
             )
         )
-    }
-
-//    @Operation(
-//        summary = "Hent enhetens ferdigstilte oppgaver",
-//        description = "Henter alle ferdigstilte oppgaver for enheten som saksbehandler har tilgang til."
-//    )
-//    @GetMapping("/enhet/{enhetId}/oppgaver/tildelte/ferdigstilte", produces = ["application/json"])
-//    fun getEnhetensFerdigstilteOppgaver(
-//        @Parameter(name = "EnhetId til enheten den ansatte jobber i")
-//        @PathVariable enhetId: String,
-//        queryParams: EnhetensFerdigstilteOppgaverQueryParams
-//    ): BehandlingerListRespons {
-//        logger.debug("Params: {}", queryParams)
-//        validateRettigheterForEnhetensTildelteOppgaver()
-//
-//        val searchCriteria = behandlingerSearchCriteriaMapper.toEnhetensFerdigstilteOppgaverSearchCriteria(
-//            enhetId = enhetId,
-//            queryParams = queryParams
-//        )
-//
-//        val esResponse = elasticsearchService.findEnhetensFerdigstilteOppgaverByCriteria(searchCriteria)
-//        return BehandlingerListRespons(
-//            antallTreffTotalt = esResponse.totalHits.toInt(),
-//            behandlinger = behandlingListMapper.mapAnonymeEsBehandlingerToListView(
-//                esBehandlinger = esResponse.searchHits.map { it.content },
-//            ),
-//        )
-//    }
-
-//    @Operation(
-//        summary = "Hent enhetens oppgaver på vent",
-//        description = "Henter alle oppgaver satt på vent for enheten som saksbehandler har tilgang til."
-//    )
-//    @GetMapping("/enhet/{enhetId}/oppgaver/tildelte/paavent", produces = ["application/json"])
-//    fun getEnhetensOppgaverPaaVent(
-//        @Parameter(name = "EnhetId til enheten den ansatte jobber i")
-//        @PathVariable enhetId: String,
-//        queryParams: EnhetensOppgaverPaaVentQueryParams
-//    ): BehandlingerListRespons {
-//        logger.debug("Params: {}", queryParams)
-//        validateRettigheterForEnhetensTildelteOppgaver()
-//
-//        val searchCriteria = behandlingerSearchCriteriaMapper.toEnhetensOppgaverPaaVentSearchCriteria(
-//            enhetId = enhetId,
-//            queryParams = queryParams
-//        )
-//
-//        val esResponse = elasticsearchService.findEnhetensOppgaverPaaVentByCriteria(searchCriteria)
-//        return BehandlingerListRespons(
-//            antallTreffTotalt = esResponse.totalHits.toInt(),
-//            behandlinger = behandlingListMapper.mapAnonymeEsBehandlingerToListView(
-//                esBehandlinger = esResponse.searchHits.map { it.content },
-//            ),
-//        )
-//    }
-
-//    @Operation(
-//        summary = "Hent uferdige oppgaver for en enhet",
-//        description = "Henter alle uferdige oppgaver i enheten som saksbehandler har tilgang til."
-//    )
-//    @GetMapping("/enhet/{enhetId}/oppgaver/tildelte/uferdige", produces = ["application/json"])
-//    fun getEnhetensUferdigeOppgaver(
-//        @Parameter(name = "EnhetId til enheten den ansatte jobber i")
-//        @PathVariable enhetId: String,
-//        queryParams: EnhetensUferdigeOppgaverQueryParams
-//    ): BehandlingerListRespons {
-//        logger.debug("Params: {}", queryParams)
-//        validateRettigheterForEnhetensTildelteOppgaver()
-//
-//        val searchCriteria = behandlingerSearchCriteriaMapper.toEnhetensUferdigeOppgaverSearchCriteria(
-//            enhetId = enhetId,
-//            queryParams = queryParams
-//        )
-//
-//        val esResponse = elasticsearchService.findEnhetensUferdigeOppgaverByCriteria(searchCriteria)
-//        return BehandlingerListRespons(
-//            antallTreffTotalt = esResponse.totalHits.toInt(),
-//            behandlinger = behandlingListMapper.mapAnonymeEsBehandlingerToListView(
-//                esBehandlinger = esResponse.searchHits.map { it.content },
-//            ),
-//        )
-//    }
-
-    private fun validateNavIdent(navIdent: String) {
-        val innloggetIdent = oAuthTokenService.getInnloggetIdent()
-        if (innloggetIdent != navIdent) {
-            throw NotMatchingUserException(
-                "logged in user does not match sent in user. " +
-                        "Logged in: $innloggetIdent, sent in: $navIdent"
-            )
-        }
     }
 
     private fun validateRettigheterForEnhetensTildelteOppgaver() {
@@ -289,14 +312,13 @@ class OppgaverListController(
     }
 
     private fun getYtelserQueryListForSaksbehandler(
-        queryParams: CommonOppgaverQueryParams,
+        ytelser: List<String>,
         tildelteYtelser: List<Ytelse>
     ): List<String> {
-        return if (queryParams.ytelser.isEmpty()) {
+        return if (ytelser.isEmpty()) {
             tildelteYtelser.map { it.id }
         } else {
-            tildelteYtelser.map { it.id }.intersect(queryParams.ytelser.toSet())
+            tildelteYtelser.map { it.id }.intersect(ytelser.toSet())
         }.toList()
     }
 }
-
