--- src/main/kotlin/no/nav/klage/search/api/view/BehandlingListViewNew.kt
@@ -1,10 +1,0 @@
-package no.nav.klage.search.api.view
-
-data class BehandlingerListResponsNew(
-    val antallTreffTotalt: Int,
-    val behandlinger: List<BehandlingListViewNew>
-)
-
-data class BehandlingListViewNew(
-    val id: String,
-)
