+++ src/main/kotlin/no/nav/klage/search/api/view/SearchResponse.kt
@@ -6,9 +6,9 @@ package no.nav.klage.search.api.view
 data class FnrSearchResponse(
     val fnr: String,
     val navn: NavnView,
-    val behandlinger: List<BehandlingListView> = emptyList(),
-    val aapneBehandlinger: List<BehandlingListView> = emptyList(),
-    val avsluttedeBehandlinger: List<BehandlingListView> = emptyList()
+    val behandlinger: List<BehandlingView> = emptyList(),
+    val aapneBehandlinger: List<BehandlingView> = emptyList(),
+    val avsluttedeBehandlinger: List<BehandlingView> = emptyList()
 )
 
 /**
