--- src/main/kotlin/no/nav/klage/search/domain/kodeverk/YtelseTilHjemler.kt
@@ -1,115 +1,0 @@
-package no.nav.klage.search.domain.kodeverk
-
-import no.nav.klage.kodeverk.Ytelse
-import no.nav.klage.kodeverk.hjemmel.Hjemmel
-import no.nav.klage.kodeverk.hjemmel.Hjemmel.*
-
-//TODO: Denne bør nok flyttes til felles kodeverk
-val ytelseTilSoekehjemler: Map<Ytelse, List<Hjemmel>> = mapOf(
-    Ytelse.OMS_PSB to listOf(
-        FTRL_9_2,
-        FTRL_9_3,
-        FTRL_9_10,
-        FTRL_9_11,
-        FTRL_9_13,
-        FTRL_9_15,
-        FTRL_9_16,
-        FTRL_8_13,
-        FTRL_22_3,
-        FTRL_22_13,
-        FTRL_22_15,
-        MANGLER,
-    ),
-    Ytelse.OMS_OLP to listOf(
-        FTRL_9_2,
-        FTRL_9_3,
-        FTRL_9_14,
-        FTRL_9_15,
-        FTRL_9_16,
-        FTRL_22_3,
-        FTRL_22_13,
-        FTRL_22_15,
-        MANGLER,
-    ),
-    Ytelse.OMS_OMP to listOf(
-        FTRL_9_2,
-        FTRL_9_3,
-        FTRL_9_5,
-        FTRL_9_6,
-        FTRL_9_8,
-        FTRL_9_9,
-        FTRL_22_3,
-        FTRL_22_13,
-        FTRL_22_15,
-        MANGLER,
-    ),
-    Ytelse.OMS_PLS to listOf(
-        FTRL_9_2,
-        FTRL_9_3,
-        FTRL_9_13,
-        FTRL_9_15,
-        FTRL_9_16,
-        FTRL_22_3,
-        FTRL_22_13,
-        FTRL_22_15,
-        MANGLER,
-    ),
-    Ytelse.SYK_SYK to listOf(
-        FTRL_8_2,
-        FTRL_8_3,
-        FTRL_8_4,
-        FTRL_8_7,
-        FTRL_8_8,
-        FTRL_8_12,
-        FTRL_8_13,
-        FTRL_8_15,
-        FTRL_8_20,
-        FTRL_8_28,
-        FTRL_8_29,
-        FTRL_8_30,
-        FTRL_8_34,
-        FTRL_8_35,
-        FTRL_8_36,
-        FTRL_8_37,
-        FTRL_8_38,
-        FTRL_8_39,
-        FTRL_8_40,
-        FTRL_8_41,
-        FTRL_8_42,
-        FTRL_8_43,
-        FTRL_8_47,
-        FTRL_8_49,
-        FTRL_22_3,
-        FTRL_22_12,
-        FTRL_22_13,
-        FTRL_22_15,
-        MANGLER,
-    ),
-    Ytelse.SUP_PEN to listOf(
-        SUP_ST_L_3,
-        SUP_ST_L_4,
-        SUP_ST_L_5,
-        SUP_ST_L_6,
-        SUP_ST_L_7,
-        SUP_ST_L_8,
-        SUP_ST_L_9,
-        SUP_ST_L_10,
-        SUP_ST_L_11,
-        SUP_ST_L_12,
-        SUP_ST_L_13,
-        MANGLER,
-    ),
-    Ytelse.SUP_UFF to listOf(
-        SUP_ST_L_3,
-        SUP_ST_L_4,
-        SUP_ST_L_5,
-        SUP_ST_L_6,
-        SUP_ST_L_8,
-        SUP_ST_L_9,
-        SUP_ST_L_10,
-        SUP_ST_L_12,
-        SUP_ST_L_13,
-        SUP_ST_L_18,
-        MANGLER,
-    ),
-)
